+++ keg/tests/test_db.py
@@ -5,6 +5,7 @@ from keg.db import db
 
 import keg_apps.db.model.entities as ents
 from keg_apps.db.app import DBApp
+from keg_apps.db2 import DB2App
 
 
 def setup_module(module):
@@ -33,3 +34,12 @@ class TestDB(object):
         db.session.add(dud)
         db.session.commit()
         assert ents.PGDud2.query.count() == 1
+
+
+class TestDB2(object):
+
+    def test_init_without_db_binds(self):
+        # Make sure we don't get an error initializing the app when the SQLALCHEMY_BINDS config
+        # option is None
+        app = DB2App.testing_prep()
+        assert app.config.get('SQLALCHEMY_BINDS') is None