+++ keg_apps/db2.py
@@ -1,0 +1,13 @@
+from __future__ import absolute_import
+from __future__ import unicode_literals
+from __future__ import print_function
+
+from keg.app import Keg
+
+
+class DB2App(Keg):
+    import_name = __name__
+    # Silence the keyring warning messages
+    keyring_enabled = False
+    db_enabled = True
+    db_visit_modules = []